backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0
python-multipart>=0.0.5
langchain>=0.1.0
langchain-community>=0.0.1
langchain-openai>=0.0.1
openai>=1.0.0
qdrant-client>=1.6.0
python-dotenv>=0.19.0
PyPDF2>=3.0.0
tiktoken>=0.5.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
tenacity>=8.0.1 